backend/requirements.txt
# FastAPI e Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Agno (Phidata) - Hub de Agentes
agno==0.2.0

# LLM e Embeddings
google-generativeai==0.3.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.10

# Processamento de Documentos
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
openpyxl==3.10.10
python-magic==0.4.27

# Banco de Dados e Vetorial
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
alembic==1.12.1

# Armazenamento
minio==7.2.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.105.0

# Autenticação e Segurança
PyJWT==2.8.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Busca e Ranking
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
logging-config==0.5.0

# Desenvolvimento e Testes
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Logging e Monitoramento
python-json-logger==2.0.7
